+++ common/src/main/java/com/viaversion/viaversion/protocols/protocol1_9to1_8/storage/EntityTracker1_9.java
@@ -152,7 +152,7 @@ public class EntityTracker1_9 extends EntityTrackerBase {
     }
 
     public boolean interactedBlockRecently(int x, int y, int z) {
-        return blockInteractions.contains(new Position(x, (short) y, z));
+        return blockInteractions.contains(new Position(x, y, z));
     }
 
     public void addBlockInteraction(Position p) {
